+++ src/main/java/duke/Storage.java
@@ -1,3 +1,9 @@
+package duke;
+
+import duke.task.Deadline;
+import duke.task.Task;
+import duke.task.ToDo;
+import duke.task.Event;
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.IOException;
