+++ src/main/java/duke/TaskMaster.java
@@ -1,3 +1,7 @@
+package duke;
+
+import duke.task.Task;
+
 import java.util.ArrayList;
 
 public class TaskMaster {
