+++ src/main/java/duke/command/AddCommand.java
@@ -1,3 +1,13 @@
+package duke.command;
+
+import duke.Storage;
+import duke.TaskMaster;
+import duke.Ui;
+import duke.task.Deadline;
+import duke.task.Event;
+import duke.task.Task;
+import duke.task.ToDo;
+
 import java.time.LocalDate;
 
 public class AddCommand extends Command {
